PVP-168 Moved ExerciseInfoUtils into a service ExerciseService
